Determine SQL parser API (#7954)
* Refactor SQLParserEngine

* Determine SQL parser API
