--- shardingsphere-sql-parser/shardingsphere-sql-parser-engine/src/main/java/org/apache/shardingsphere/sql/parser/api/visitor/SQLVisitorEngineFactory.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.sql.parser.api.visitor;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-/**
- * SQL visitor engine factory.
- */
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class SQLVisitorEngineFactory {
-    
-    /**
-     * Get SQL visitor engine.
-     *
-     * @param databaseType database type
-     * @param visitorType visitor type
-     * @param <T> type of visitor result
-     * @return SQL visitor engine
-     */
-    public static <T> SQLVisitorEngine<T> getSQLVisitorEngine(final String databaseType, final String visitorType) {
-        return new SQLVisitorEngine<>(databaseType, visitorType);
-    }
-}
